Big light and more update
Added lights to emissive mushies in the forest

Generated lightmap UV's for all the tree's and foliage in forest so they can benefit from mushie lights

Enabled and increased lightmap size on terrain to "1" so it can benefit from pretty lights

Decided on baked lights for the church candles. The flicker doesn't translate to baked lights and them flickering on and off just looks bad.

Kept flicker on braziers in graveyard and chimney fire in the meadow and went with "mixed" lighting. I don't believe the light limit will effect these.

Ticked "Contribute global illumination for many objects such as church, mausoleum, tunnel, tunnel ramps, tunnel doors, cave and meadow house among others so it can benefit from pretty lights.

Increased church candle intensity from "5" to "10" looks fine but may go back to 5.

Added box collider to podium in church.

Changed all lights to be the same color green.

Upon playtesting I decided to with an intensity of "5" for the church candles.

Added "Inside sounds" for the cave, mausoleums and church. Made a simple script that plays sounds when player is inside the box collider and doesn't when he/she is outside of it. Runs smooth.

Playtested everything looks good, runs good it's good! <('O')>