SceneYUpdate
Fixed scaling for entire bathroom making it feel a lot more open.

Fixed transforms for ship.

Imported URPwater but will have to fine tune settings.

Added a slight "BathroomSteam effect for the entire bathroom.

Added "stormyclouds" that have a slight steam effect to them to go with the theme of the level.

Added lighting to the stormyclouds.

Imported a door for the bathroom.